refactor: redis code has its own package
